Move component and feature injections into utils
